[QNN EP] Fix comment issues
- Move custom-op files into a new custom_op/ folder
- Extract the env-var-read + domain-build logic out of QnnEpFactory's
  constructor into BuildCustomOpDomainsFromEnv() in the new
  custom_op/qnn_custom_op_domain_registry
- Document InferOutputShape's FLOAT-only assumption as a known,
  fail-fast limitation
- Fix some minor issues
