+++ docs/execution_providers/QNN-ExecutionProvider.md
@@ -1738,7 +1738,10 @@ export ORT_QNN_CUSTOM_OP_DOMAINS="udo_domain:MyAdd;other_domain:OpA,OpB"
 > **Note:** `ORT_QNN_CUSTOM_OP_DOMAINS` must be set before `libonnxruntime_providers_qnn.so` is
 > loaded (i.e., before the ORT environment is created). The env var names the *domain and op-types*
 > needed for model-load schema validation. The actual QNN HW kernel still comes from `op_packages` at
-> session time; the two config items are complementary.
+> session time; the two config items are complementary. The registered placeholder schema currently
+> assumes FLOAT input/output element types (matching all supported UDO usage today); a UDO whose
+> ONNX-level type is genuinely non-FLOAT will fail session construction with a clear type-inference
+> error rather than silently misbehaving.
 
 > **Optional — manual domain registration for CPU fallback:** If you need a real CPU kernel to run
 > the op on the CPU EP (e.g., for accuracy comparison), you can still construct an
